ps3: Move wave parser out of own folder
